Raw message sanity check
When grabbing a numeric, check to make sure the json element is indeed a number. During testing, discovered that crypto messages of one particular exchange (kucoin) arrive in the json array with the numeric values quoted. Without this added check, the attempt to grab a double throws. I sent an email to Tiingo explaining the inconsistency; maybe they can fix it.